feat: move key analytics SCAN logic into DatabasePort adapter for agent support
