+++ packages/agent/src/command-executor.ts
@@ -1,4 +1,6 @@
 import Valkey from 'iovalkey';
+import type { KeyAnalyticsOptions, KeyPatternData } from '@betterdb/shared';
+import { extractPattern } from '@betterdb/shared';
 
 const ALLOWED_COMMANDS: Set<string> = new Set([
   'PING', 'INFO', 'DBSIZE',
@@ -12,6 +14,7 @@ const ALLOWED_COMMANDS: Set<string> = new Set([
   'COMMAND',
   'ROLE',
   'LASTSAVE',
+  'COLLECT_KEY_ANALYTICS',
 ]);
 
 // Subcommands that are explicitly allowed for each command
@@ -79,8 +82,95 @@ export class CommandExecutor {
       return this.client.config('GET', ...args.slice(1));
     }
 
+    if (upperCmd === 'COLLECT_KEY_ANALYTICS' && args) {
+      return this.executeCollectKeyAnalytics(args[0]);
+    }
+
     // For all other commands, use call()
     const callArgs = args ? [upperCmd, ...args] : [upperCmd];
     return this.client.call(...(callArgs as [string, ...string[]]));
+  }
+
+  private async executeCollectKeyAnalytics(optionsJson: string): Promise<string> {
+    const options: KeyAnalyticsOptions = JSON.parse(optionsJson);
+    const dbSize = await this.client.dbsize();
+
+    if (dbSize === 0) {
+      return JSON.stringify({ dbSize: 0, scanned: 0, patterns: [] });
+    }
+
+    const patternsMap = new Map<string, KeyPatternData>();
+    let cursor = '0';
+    let scanned = 0;
+
+    do {
+      const [newCursor, keys] = await this.client.scan(cursor, 'COUNT', options.scanBatchSize);
+      cursor = newCursor;
+
+      for (const key of keys) {
+        if (scanned >= options.sampleSize) break;
+        scanned++;
+
+        const pattern = extractPattern(key);
+        const stats = patternsMap.get(pattern) || {
+          pattern,
+          count: 0,
+          totalMemory: 0,
+          maxMemory: 0,
+          totalIdleTime: 0,
+          withTtl: 0,
+          withoutTtl: 0,
+          ttlValues: [],
+          accessFrequencies: [],
+        };
+
+        try {
+          const pipeline = this.client.pipeline();
+          pipeline.memory('USAGE', key);
+          pipeline.object('IDLETIME', key);
+          pipeline.object('FREQ', key);
+          pipeline.ttl(key);
+
+          const results = (await pipeline.exec()) || [];
+          const [memResult, idleResult, freqResult, ttlResult] = results;
+
+          stats.count++;
+
+          if (memResult && memResult[1] !== null) {
+            const mem = memResult[1] as number;
+            stats.totalMemory += mem;
+            if (mem > stats.maxMemory) stats.maxMemory = mem;
+          }
+
+          if (idleResult && idleResult[1] !== null) {
+            stats.totalIdleTime += idleResult[1] as number;
+          }
+
+          if (freqResult && freqResult[1] !== null) {
+            stats.accessFrequencies.push(freqResult[1] as number);
+          }
+
+          const ttl = ttlResult?.[1] as number;
+          if (ttl > 0) {
+            stats.withTtl++;
+            stats.ttlValues.push(ttl);
+          } else {
+            stats.withoutTtl++;
+          }
+
+          patternsMap.set(pattern, stats);
+        } catch {
+          // Skip keys that can't be inspected
+        }
+      }
+
+      if (scanned >= options.sampleSize) break;
+    } while (cursor !== '0');
+
+    return JSON.stringify({
+      dbSize,
+      scanned,
+      patterns: Array.from(patternsMap.values()),
+    });
   }
 }
